Changed Create Parameter Layout method
Changed the createParameterLayout() method to use the std::vector method of creating parameters as seen in The Audio Programmer Youtube Video: https://www.youtube.com/watch?v=nkQPsYOdIrk . In addition, to reflect the changes made in previous commit, the params and GetParams enum and function were replaced by the X_ID and X_NAME where X is the respective parameter. Please note that we are only implementing the Compressor for one full range band (we are using without loss of generality the low band) as of this commit.